add check for ProductExistence to deleteProductById

+++ backend/src/main/java/com/github/hh/backend/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.github.hh.backend.service;
 
+import com.github.hh.backend.exception.NoSuchProductException;
 import com.github.hh.backend.model.Product;
 import com.github.hh.backend.model.ProductDTO;
 import com.github.hh.backend.repository.ProductRepo;
@@ -32,6 +33,9 @@ public class ProductService {
     }
 
     public void deleteProductById(String id) {
+        if (!productRepo.existsById(id)) {
+            throw new NoSuchProductException("Product with ID " + id + " does not exist");
+        }
         productRepo.deleteById(id);
     }
 }
